requirements.txt
Flask==2.3.3
requests==2.31.0
pathlib==1.0.1
google-generativeai